requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
mysql==0.0.2
mysql-connector-python==8.0.22
mysqlclient==1.4.6
oauthlib==3.1.0
praw==6.5.1
prawcore==1.0.1
PySocks==1.7.1
python-dotenv==0.12.0
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
tweepy==3.8.0
update-checker==0.16
urllib3==1.25.8
websocket-client==0.57.0
